+++ src/main/java/net/parkl/ocpp/service/cs/cleanup/TransactionCleanupService.java
@@ -79,6 +79,6 @@ public class TransactionCleanupService {
 
     @Transactional
     public int cleanupConsumptionStates(Date date) {
-        return (int) chargingConsumptionStateRepo.deleteBefore(date);
+        return chargingConsumptionStateRepo.deleteBefore(date);
     }
 }
